photo-classifier/requirements.txt
absl-py==2.3.0
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.58.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.72.1
h5py==3.13.0
idna==3.10
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
markdown==3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
numpy==1.26.4
opt-einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
protobuf==4.25.8
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing-extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
